fix: 리뷰 반영 - self-injection 제거, CompletePaymentUseCase 추출
PendingPaymentProcessor의 @DistributedLock+@Transactional 메서드를
별도 CompletePaymentUseCase로 추출하여 AOP self-invocation 문제 해결
