+++ SClass-Batch/src/test/kotlin/com/sclass/batch/payment/PendingPaymentProcessorTest.kt
@@ -23,6 +23,7 @@ class PendingPaymentProcessorTest {
     private lateinit var productAdaptor: ProductAdaptor
     private lateinit var coinDomainService: CoinDomainService
     private lateinit var pgGateway: PgGateway
+    private lateinit var completePaymentUseCase: CompletePaymentUseCase
     private lateinit var processor: PendingPaymentProcessor
 
     @BeforeEach
@@ -31,9 +32,8 @@ class PendingPaymentProcessorTest {
         productAdaptor = mockk()
         coinDomainService = mockk()
         pgGateway = mockk()
-        processor = PendingPaymentProcessor(paymentAdaptor, productAdaptor, coinDomainService, pgGateway, mockk())
-        // self-injection: 단위테스트에서는 프록시 없이 자기 자신을 주입
-        processor = PendingPaymentProcessor(paymentAdaptor, productAdaptor, coinDomainService, pgGateway, processor)
+        completePaymentUseCase = CompletePaymentUseCase(paymentAdaptor, productAdaptor, coinDomainService)
+        processor = PendingPaymentProcessor(pgGateway, completePaymentUseCase)
     }
 
     private fun createPendingPayment() =
